+++ src/index.html
@@ -19,7 +19,7 @@
             <button class="enter-icon"></button>
         </form>
         <div id="todo-list" class="todo-list"></div>
-        <button class="clear-all">
+        <button class="clear-all" id="clear-all-completed">
             Clear all completed
         </button>
     </main>
