+++ src/style.css
@@ -134,3 +134,7 @@ form input {
   opacity: 0.5;
   display: block;
 }
+
+input[type=checkbox]:checked+.list {
+  text-decoration: line-through;
+}
